hotfix: location prefix issues
We were using the location ID rather than CODE as was intended. The tests were not defining usable codes, so I updated that.
The location length detection also didn't filter on valid records nor on the Location type (district)
Improved testing with different lengths too.
This is meant to be applied as a hotfix in Niger